Verifier test with foundry
